requirements.txt
alpaca-trade-api==4.0.0
flask
pandas
numpy
python-dotenv
yfinance
matplotlib
optuna
scipy
